+++ packages/expo-router/CHANGELOG.md
@@ -6,6 +6,7 @@
 
 ### 🎉 New features
 
+- add `<Stack.Toolbar.Link>` component for iOS 26 zoom transition.
 - improve assignment of Stack.Header props to Stack options ([#42544](https://github.com/expo/expo/pull/42544) by [@Ubax](https://github.com/Ubax))
 - fallback to title when label is not present in Stack.Toolbar.Menu ([#42531](https://github.com/expo/expo/pull/42531) by [@Ubax](https://github.com/Ubax))
 - [ios] add iconRenderingMode to toolbar items ([#42346](https://github.com/expo/expo/pull/42346) by [@Ubax](https://github.com/Ubax))